requirements.txt
spotipy==2.13.0
mysqlclient==1.4.6
django-debug-toolbar==2.2
